implemented a minimal ui
As part of the revisit of the original quartz parchment shears,
we were instructed to use our newly acquired knowledge in dom
manipulation to make our game more interactive.
I have replaced the clunky prompt window for user input with buttons.
The results for each round, as well as the verdict for a particular game,
are now displayed within the browser window, instead of the console.